Added xstatus_draw_text_widget() to unify text rendering.  Used within status_file.   Unified return point in draw_status_file().   Unified return point in format() in temperature module.  Used basic types within XSTextWidget.  Removed font_size field.

+++ status_file.h
@@ -4,5 +4,5 @@
 #include <stdint.h>
 #include <xcb/xcb.h>
 // Returns offset for next widget
-int draw_status_file(xcb_connection_t * xc, int x, const char * filename);
+int draw_status_file(xcb_connection_t * xc, short x, const char * filename);
 #endif//!XSTATUS_STATUS_FILE_H
